Code refactored.

+++ Rendering/Controls/OpenGL/Colorado.Rendering.Controls.OpenGL.OpenGLAPI/Wrappers/Rendering/IFastRenderingData.cs
@@ -7,7 +7,7 @@ namespace Colorado.Rendering.Controls.OpenGL.OpenGLAPI.Wrappers.Rendering
         double[] VerticesValuesArray { get; }
         double[] NormalsValuesArray { get; }
         int VerticesCount { get; }
-        double[] VerticesColorsValuesArray { get; }
+        byte[] VerticesColorsValuesArray { get; }
 
         Primitive Primitive { get; }
     }
